Removing .choices() logic from muody cli.

+++ src/config/hostsFunctions.ts
@@ -1,4 +1,4 @@
-import fs from 'fs';
+import fs, { exists } from 'fs';
 import { Map } from 'immutable';
 import util from 'util';
 
@@ -41,18 +41,33 @@ export const createIHostObject = (args: IArgs): IHostData => {
 };
 
 export const handleJSONFile = async (command: string, args: IArgs): Promise<Map<string, IHostData>> => {
+    // TODO: error handling
     const hostFile: IHosts = await showHostsContent();
     const mapHostFile: Map<string, IHostData> = Map(hostFile);
     const newHostToAdd: IHostData = createIHostObject(args);
 
     switch (command) {
         case 'add':
+            if (!mapHostFile.has(args.host)) {
+                const changedHosts: Map<string, IHostData> = mapHostFile.set(args.host, newHostToAdd);
+                return Promise.resolve(changedHosts);
+            }
+
+            return Promise.reject('This profile already exists. Try muody update --host=...');
         case 'update':
-            const changedHosts: Map<string, IHostData> = mapHostFile.set(args.host, newHostToAdd);
-            return Promise.resolve(changedHosts);
+            if (mapHostFile.has(args.host)) {
+                const changedHosts: Map<string, IHostData> = mapHostFile.set(args.host, newHostToAdd);
+                return Promise.resolve(changedHosts);
+            }
+
+            return Promise.reject("Can't update profile that doesn't exists.");
         case 'delete':
-            const reducedHosts: Map<string, IHostData> = mapHostFile.delete(args.host);
-            return Promise.resolve(reducedHosts);
+            if (mapHostFile.has(args.host)) {
+                const reducedHosts: Map<string, IHostData> = mapHostFile.delete(args.host);
+                return Promise.resolve(reducedHosts);
+            }
+
+            return Promise.reject("This host doesn't exists.");
     }
 
 };
